Move log server response to system monitoring
